boards: shields: adafruit_adalogger_featherwing: add pcf8523 rtc
Add support for the PCF8523 I2C RTC included on the Adafruit Adalogger
featherwing.

This work is based on the Adafruit Data Logger overlay by Henrik Brix
Andersen [1]. The main difference is that the interrupt pin of the
pcf8523 RTC is not connected to any IO on the Adalogger Featherwing.

[1]:
https://github.com/zephyrproject-rtos/zephyr/blob/main/boards/shields/adafruit_data_logger/adafruit_data_logger.overlay